docker/requirements.txt
streamlit
transformers
torch
Pillow
numpy
opencv-python-headless
scipy
streamlit-webrtc
av